# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fact Checker Application API
  description: |
    A comprehensive fact-checking platform API that provides endpoints for user authentication, 
    fact management, community interaction, moderation, and administrative functions.
    
    ## Features
    - User authentication and profile management
    - Fact submission and verification
    - Community voting and commenting
    - Content moderation and reporting
    - Analytics and insights
    - Administrative dashboard
    - UI framework components
    
    ## Authentication
    This API uses session-based authentication. Users must log in to access protected endpoints.
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@factchecker.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.factchecker.com
    description: Production server
...
paths:
...
  /comments:
    post:
      tags: [Comments]
      summary: Create comment
      description: Post a new comment on a fact
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, fact_id]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 1000
                fact_id:
                  type: string
                  format: uuid
                parent_id:
                  type: string
                  format: uuid
                  nullable: true
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      comment:
                        $ref: '#/components/schemas/Comment'
  /comments/fact/{fact_id}:
    get:
      tags: [Comments]
      summary: Get fact comments
      description: Retrieve comments for a specific fact
      parameters:
        - name: fact_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, most_voted]
            default: newest
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      comments:
                        type: array
                        items:
                          $ref: '#/components/schemas/CommentWithReplies'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

  # Admin Endpoints
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        user_id:
          type: string
          format: uuid
        fact_id:
          type: string
          format: uuid
        parent_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        vote_counts:
          $ref: '#/components/schemas/CommentVoteCounts'
    CommentWithReplies:
      allOf:
        - $ref: '#/components/schemas/Comment'
        - type: object
          properties:
            replies:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
...
    CommentVoteCounts:
      type: object
      properties:
        upvotes:
          type: integer
        downvotes:
          type: integer
    Pagination:
      type: object
      properties:
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
security:
  - sessionAuth: []